Add support for element attributes, numbering formats
This adds support in the style grammar (and the corresponding reader
code) for arbitrary attributes to be added to an output element, using
the format `p{a=b,c=d}` where `a` and `c` are attribute names and `b`
and `d` are attribute values. It also adds a specific use-case in the
form of supporting different formats for numbered lists (e.g. letters,
Roman numerals).